+++ packages/web3-validator/src/validation/bytes.ts
@@ -22,7 +22,8 @@ import { isHexStrict } from './string.js';
 /**
  * checks input if typeof data is valid Uint8Array input
  */
-export const isUint8Array = (data: ValidInputTypes) => data instanceof Uint8Array;
+export const isUint8Array = (data: ValidInputTypes): data is Uint8Array =>
+	data instanceof Uint8Array || data?.constructor?.name === 'Uint8Array';
 
 export const isBytes = (
 	value: ValidInputTypes | Uint8Array | number[],
@@ -30,7 +31,7 @@ export const isBytes = (
 		abiType: 'bytes',
 	},
 ) => {
-	if (typeof value !== 'string' && !Array.isArray(value) && !(value instanceof Uint8Array)) {
+	if (typeof value !== 'string' && !Array.isArray(value) && !isUint8Array(value)) {
 		return false;
 	}