+++ packages/web3-eth-accounts/CHANGELOG.md
@@ -154,3 +154,4 @@ Documentation:
 ### Fixed
 
 -   Send Transaction config used to be ignored if the passed `common` did not have a `copy()` and the `chainId` was not provided (#6663)
+-   Fixed an issue with detecting Uint8Array (#6486)